+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/HypixelModApiEvents.kt
@@ -24,8 +24,10 @@ object HypixelModApiEvents : RegisteredEvent {
 
     object LocationEventManager : AbstractEventManager<(ClientboundLocationPacket) -> Unit, LocationEventManager.LocationEvent>() {
         fun runTasks(packet: ClientboundLocationPacket) {
-            for (task in tasks) {
-                task.callback(packet)
+            safeExecution {
+                for (task in tasks) {
+                    task.callback(packet)
+                }
             }
         }